backend/requirements.txt
async-timeout==4.0.2
beautifulsoup4==4.11.1
certifi==2022.6.15.1
charset-normalizer==2.1.1
Deprecated==1.2.13
idna==3.3
lxml==4.9.1
packaging==21.3
pyparsing==3.0.9
python-dotenv==0.21.0
redis==4.3.4
requests==2.28.1
soupsieve==2.3.2.post1
urllib3==1.26.12
wrapt==1.14.1
